Final formatting and documentation
