+++ Readme.md
@@ -1,21 +1,15 @@
-# Qiagen Upload
+# Oncodeep Upload
 
-This repository contains the scripts required for use in uploading TSO500 samples to the OncoKDM for analysis via the SFTP server. The repository has a logger script which provides logging functionality. It contains a [Dockerfile](Dockerfile) and a [Makefile](Makefile) for use in building a dockerised version and pushing this dockerised version to Docker Hub. It also contains an [XML template](templates/sample_upload_template.xml) which is used to create the XML metadata file used for each sample upload.
+This repository contains the [oncodeep_upload script](oncodeep_upload/oncodeep_upload.py) required for use in uploading files to the OncoKDM for analysis via the SFTP server. The repository has a logger script which provides logging functionality. It contains a [Dockerfile](Dockerfile) and a [Makefile](Makefile) for use in building a dockerised version and pushing this dockerised version to Docker Hub.
 
+## Usage
 
-### [oncodeep_upload](oncodeep_upload)
-
-This module contains the [oncodeep_upload script](oncodeep_upload/oncodeep_upload.py), which is run to upload the fastqs onto the oncodeep SFTP server.
-
-#### Requirements
+### Requirements
 
 Python packages are specified in the requirements.txt file. Further dependencies are:
-* sshpass : ```sudo apt install sshpass```
 * docker
 
-#### Usage
-
-##### Inputs
+### Inputs
 
 The inputs are as follows:
 ```bash
@@ -37,23 +31,20 @@ options:
                         Password for the OncoDEEP SFTP server
 ```
 
-##### Outputs
+### Outputs
 
-The script has 4 output files:
-* Logfile - ```outputs/get_user_code_YYYYMMDD_HHMMSS.log``` - Contains log messages documenting script logic
-* Code verifier file - ```outputs/qiagen_code_verifier_YYYYMMDD_HHMMSS``` - Contains the code_verifier generated by the script. A high-entropy cryptographic random string
-* Device code file - ```outputs/qiagen_device_code_YYYYMMDD_HHMMSS``` - Contains the device_code generated by the script. This code authorises the device.
-* User code file - ```outputs/qiagen_user_code_YYYYMMDD_HHMMSS``` - Contains the user code generated by the script. This can be used to register the device in QiaOAuth
+The script has 1 output file:
+* Logfile - ```outputs/oncodeep_upload.$FILESTEM.YYYYMMDD_HHMMSS.log``` - Contains log messages documenting script logic
 
 
-##### Script
+### Script
 
 The script can be run as follows:
 ```bash
 python3 -m oncodeep_upload -R $RUN_IDENTIFIER -F $SRC_FILEPATH -H $HOSTNAME -U $USERNAME -P $PASSWORD
 ```
 
-##### Docker Image
+### Docker Image
 
 The docker image is built, tagged and saved as a .tar.gz file using the [Makefile](Makefile) as follows:
