+++ oncodeep_upload/oncodeep_upload.py
@@ -1,8 +1,10 @@
-"""oncodeep_upload.py
+#!/usr/bin/env python3
+""" oncodeep_upload.py
 
 This script uploads an input file to a specified destination folder on an SFTP server with specified
 hostname, username and password, using paramiko
 """
+
 import sys
 import os
 import paramiko
@@ -22,10 +24,10 @@ class OncoDeepUpload:
         hostname (str):                     SFTP server login hostname
         username (str):                     SFTP server login username
         password (str):                     SFTP server login password
-        sftp_client
-        (paramiko.sftp_client.SFTPClient):  SFTP client object for use in file upload
         ssh_client
         (paramiko.client.SSHClient):        SSH client object for use in file upload
+        sftp_client
+        (paramiko.sftp_client.SFTPClient):  SFTP client object for use in file upload
 
     Methods
         create_ssh_client()
@@ -51,16 +53,17 @@ class OncoDeepUpload:
         Constructor for the OncoDeepUpload class
             :param file_path (str):         File path of file to upload
             :param run_identifier (str):    Name of the run, to be used to create SFTP server subdir
-            :param subfolder (str):         Subfolder on SFTP server to load file into
-            :param logger (logging.Logger): Logger object
             :param hostname (str):          SFTP server hostname
             :param username (str):          SFTP server login username
             :param password (str):          SFTP server login password
+            :param logger (logging.Logger): Logger object
         """
         self.file_path = file_path
         self.run_identifier = run_identifier
         self.subfolder = f"/data/{self.run_identifier}/"
-        self.sftp_destination = os.path.join(self.subfolder, os.path.basename(file_path))
+        self.sftp_destination = os.path.join(
+            self.subfolder, os.path.basename(file_path)
+        )
         self.logger = logger
         # Remote server credentials
         self.hostname = hostname
@@ -113,11 +116,15 @@ class OncoDeepUpload:
             :return None:
         """
         try:
-            self.logger.info(f"Uploading file to SFTP server subfolder. Src: {self.file_path} Dest: {self.sftp_destination}")
+            self.logger.info(
+                f"Uploading file to SFTP server subfolder. Src: {self.file_path} Dest: {self.sftp_destination}"
+            )
             self.sftp_client.put(self.file_path, self.sftp_destination)
             self.logger.info("File upload successful")
         except Exception as exception:
-            self.logger.error(f"An exception was encountered when uploading the file to the SFTP server: {exception}")
+            self.logger.error(
+                f"An exception was encountered when uploading the file to the SFTP server: {exception}"
+            )
             sys.exit(1)
 
     def close_connection(self) -> None:
